# specs/001-physical-ai-robotics-book/contracts/rag_api_openapi.yaml
openapi: 3.0.0
info:
  title: RAG Chatbot API
  version: 1.0.0
  description: API for the Retrieval Augmented Generation (RAG) chatbot backend.

servers:
  - url: http://localhost:8000
    description: Local development server

paths:
  /embed:
    post:
      summary: Embed text content
      description: Takes text content and stores its embedding in the vector database.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: string
                  description: The text content to embed.
                  example: "ROS 2 is a flexible framework for writing robot software."
                metadata:
                  type: object
                  description: Optional metadata associated with the text (e.g., source, chapter).
                  example: {"source": "Module 1", "chapter": "ROS 2 Fundamentals"}
      responses:
        '200':
          description: Embedding successful.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  id:
                    type: string
                    description: The ID of the embedded document.
        '400':
          description: Invalid request.

  /query:
    post:
      summary: Query the RAG system
      description: Retrieves relevant information from the embedded book content based on a user query.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                query:
                  type: string
                  description: The user's natural language query.
                  example: "What are ROS 2 nodes?"
      responses:
        '200':
          description: Successful query.
          content:
            application/json:
              schema:
                type: object
                properties:
                  answer:
                    type: string
                    example: "ROS 2 nodes are computational processes that perform specific tasks within the ROS 2 ecosystem, such as controlling a motor or processing sensor data."
                  sources:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        text_snippet:
                          type: string
                        metadata:
                          type: object
        '400':
          description: Invalid request.

  /selected-text-query:
    post:
      summary: Query RAG with user-selected text
      description: Answers questions using only the provided user-selected text as context.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                query:
                  type: string
                  description: The user's question.
                  example: "What is the main concept discussed here?"
                selected_text:
                  type: string
                  description: The text selected by the user to use as context for the query.
                  example: "Digital twins are virtual representations of physical objects or systems."
      responses:
        '200':
          description: Successful query with selected text.
          content:
            application/json:
              schema:
                type: object
                properties:
                  answer:
                    type: string
                    example: "The main concept discussed is digital twins, which are virtual representations of physical objects or systems."
                  sources:
                    type: array
                    items:
                      type: object
                      properties:
                        text_snippet:
                          type: string
        '400':
          description: Invalid request.